CommentDao deleteComment method modified to validate id emptyness MovieDao validIdValue method modified to verify whether movieId is a valid hexadecimal string UserDao addUser method modified to execute insertOne instead of updateOne (Ticket Errors Handling)

+++ src/main/java/mflix/api/daos/CommentDao.java
@@ -115,6 +115,9 @@ public class CommentDao extends AbstractMFlixDao {
      * @return true if successful deletes the comment.
      */
     public boolean deleteComment(String commentId, String email) {
+        if (commentId.isEmpty()) {
+            throw new IllegalArgumentException("commentId can't be empty");
+        }
         DeleteResult result = null;
         try {
             result = commentCollection.deleteOne(and(
